Move WheelEventProcessingSteps to ScrollingCoordinatorTypes.h https://bugs.webkit.org/show_bug.cgi?id=249519 rdar://103471270
Reviewed by Cameron McCormack.

WheelEventProcessingSteps and WheelEventHandlingResult are going to be used in more places;
move them into ScrollingCoordinatorTypes.h.

Prepare messages.py for WheelEventProcessingSteps being used in IPC messages.

* Source/WebCore/page/scrolling/ScrollingCoordinatorTypes.cpp:
(WebCore::operator<<):
* Source/WebCore/page/scrolling/ScrollingCoordinatorTypes.h:
(WebCore::WheelEventHandlingResult::needsMainThreadProcessing const):
(WebCore::WheelEventHandlingResult::handled):
(WebCore::WheelEventHandlingResult::unhandled):
(WebCore::WheelEventHandlingResult::result):
* Source/WebCore/page/scrolling/ScrollingTree.h:
(WebCore::WheelEventHandlingResult::needsMainThreadProcessing const): Deleted.
(WebCore::WheelEventHandlingResult::handled): Deleted.
(WebCore::WheelEventHandlingResult::unhandled): Deleted.
(WebCore::WheelEventHandlingResult::result): Deleted.
* Source/WebCore/platform/PlatformWheelEvent.cpp:
* Source/WebCore/platform/PlatformWheelEvent.h:
* Source/WebKit/Scripts/webkit/messages.py:
(headers_for_type):
* Source/WebKit/WebProcess/WebPage/EventDispatcher.h:

Canonical link: https://commits.webkit.org/258037@main
